+++ src/Plugins/OctoPatch.Plugin.Keyboard/GlobalKeyboardHook.cs
@@ -1,0 +1,210 @@
+﻿using System;
+using System.ComponentModel;
+using System.Diagnostics;
+using System.Runtime.InteropServices;
+
+namespace OctoPatch.Plugin.Keyboard
+{
+    /// <summary>
+    /// Based on https://gist.github.com/Stasonix
+    /// </summary>
+    internal class GlobalKeyboardHook : IDisposable
+    {
+        public event EventHandler<GlobalKeyboardHookEventArgs> KeyboardPressed;
+
+        private delegate IntPtr HookProc(int nCode, IntPtr wParam, IntPtr lParam);
+
+        [DllImport("kernel32.dll", CharSet = CharSet.Unicode)]
+        private static extern IntPtr LoadLibrary(string lpFileName);
+
+        [DllImport("kernel32.dll", CharSet = CharSet.Unicode)]
+        private static extern bool FreeLibrary(IntPtr hModule);
+
+        /// <summary>
+        /// The SetWindowsHookEx function installs an application-defined hook procedure into a hook chain.
+        /// You would install a hook procedure to monitor the system for certain types of events. These events are
+        /// associated either with a specific thread or with all threads in the same desktop as the calling thread.
+        /// </summary>
+        /// <param name="idHook">hook type</param>
+        /// <param name="lpfn">hook procedure</param>
+        /// <param name="hMod">handle to application instance</param>
+        /// <param name="dwThreadId">thread identifier</param>
+        /// <returns>If the function succeeds, the return value is the handle to the hook procedure.</returns>
+        [DllImport("USER32", SetLastError = true)]
+        private static extern IntPtr SetWindowsHookEx(int idHook, HookProc lpfn, IntPtr hMod, int dwThreadId);
+
+        /// <summary>
+        /// The UnhookWindowsHookEx function removes a hook procedure installed in a hook chain by the SetWindowsHookEx function.
+        /// </summary>
+        /// <param name="hhk">handle to hook procedure</param>
+        /// <returns>If the function succeeds, the return value is true.</returns>
+        [DllImport("USER32", SetLastError = true)]
+        private static extern bool UnhookWindowsHookEx(IntPtr hHook);
+
+        /// <summary>
+        /// The CallNextHookEx function passes the hook information to the next hook procedure in the current hook chain.
+        /// A hook procedure can call this function either before or after processing the hook information.
+        /// </summary>
+        /// <param name="hHook">handle to current hook</param>
+        /// <param name="code">hook code passed to hook procedure</param>
+        /// <param name="wParam">value passed to hook procedure</param>
+        /// <param name="lParam">value passed to hook procedure</param>
+        /// <returns>If the function succeeds, the return value is true.</returns>
+        [DllImport("USER32", SetLastError = true)]
+        private static extern IntPtr CallNextHookEx(IntPtr hHook, int code, IntPtr wParam, IntPtr lParam);
+
+        private const int WH_KEYBOARD_LL = 13;
+
+        private IntPtr _windowsHookHandle;
+        private IntPtr _user32LibraryHandle;
+        private HookProc _hookProc;
+
+        public GlobalKeyboardHook()
+        {
+            _windowsHookHandle = IntPtr.Zero;
+            _user32LibraryHandle = IntPtr.Zero;
+
+            // we must keep alive _hookProc, because GC is not aware about SetWindowsHookEx behaviour.
+            _hookProc = LowLevelKeyboardProc;
+
+            _user32LibraryHandle = LoadLibrary("User32");
+            if (_user32LibraryHandle == IntPtr.Zero)
+            {
+                int errorCode = Marshal.GetLastWin32Error();
+                throw new Win32Exception(errorCode,
+                    "Failed to load library 'User32.dll'. " +
+                    $"Error {errorCode}: {new Win32Exception(errorCode).Message}.");
+            }
+
+            _windowsHookHandle = SetWindowsHookEx(WH_KEYBOARD_LL, _hookProc, _user32LibraryHandle, 0);
+            if (_windowsHookHandle == IntPtr.Zero)
+            {
+                int errorCode = Marshal.GetLastWin32Error();
+                throw new Win32Exception(errorCode,
+                    $"Failed to adjust keyboard hooks for '{Process.GetCurrentProcess().ProcessName}'. " +
+                    $"Error {errorCode}: {new Win32Exception(errorCode).Message}.");
+            }
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (disposing)
+            {
+                // because we can unhook only in the same thread, not in garbage collector thread
+                if (_windowsHookHandle != IntPtr.Zero)
+                {
+                    if (!UnhookWindowsHookEx(_windowsHookHandle))
+                    {
+                        int errorCode = Marshal.GetLastWin32Error();
+                        throw new Win32Exception(errorCode,
+                            $"Failed to remove keyboard hooks for '{Process.GetCurrentProcess().ProcessName}'. " +
+                            $"Error {errorCode}: {new Win32Exception(errorCode).Message}.");
+                    }
+
+                    _windowsHookHandle = IntPtr.Zero;
+
+                    // ReSharper disable once DelegateSubtraction
+                    _hookProc -= LowLevelKeyboardProc;
+                }
+            }
+
+            if (_user32LibraryHandle != IntPtr.Zero)
+            {
+                // reduces reference to library by 1.
+                if (!FreeLibrary(_user32LibraryHandle))
+                {
+                    int errorCode = Marshal.GetLastWin32Error();
+                    throw new Win32Exception(errorCode,
+                        $"Failed to unload library 'User32.dll'. " +
+                        $"Error {errorCode}: {new Win32Exception(errorCode).Message}.");
+                }
+
+                _user32LibraryHandle = IntPtr.Zero;
+            }
+        }
+
+        ~GlobalKeyboardHook()
+        {
+            Dispose(false);
+        }
+
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        private IntPtr LowLevelKeyboardProc(int nCode, IntPtr wParam, IntPtr lParam)
+        {
+            var handled = false;
+            var wparamTyped = wParam.ToInt32();
+
+            if (Enum.IsDefined(typeof(KeyboardState), wparamTyped))
+            {
+                var keyboardEvent = Marshal.PtrToStructure<LowLevelKeyboardInputEvent>(lParam);
+                var eventArguments = new GlobalKeyboardHookEventArgs(keyboardEvent, (KeyboardState)wparamTyped);
+
+                KeyboardPressed?.Invoke(this, eventArguments);
+
+                handled = eventArguments.Handled;
+            }
+
+            return handled ? (IntPtr)1 : CallNextHookEx(IntPtr.Zero, nCode, wParam, lParam);
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct LowLevelKeyboardInputEvent
+        {
+            /// <summary>
+            /// A virtual-key code. The code must be a value in the range 1 to 254.
+            /// </summary>
+            public int VirtualCode;
+
+            /// <summary>
+            /// A hardware scan code for the key.
+            /// </summary>
+            public int HardwareScanCode;
+
+            /// <summary>
+            /// The extended-key flag, event-injected Flags, context code, and transition-state flag.
+            /// This member is specified as follows. An application can use the following values
+            /// to test the keystroke Flags. Testing LLKHF_INJECTED (bit 4) will tell you whether
+            /// the event was injected. If it was, then testing LLKHF_LOWER_IL_INJECTED (bit 1) will tell
+            /// you whether or not the event was injected from a process running at lower integrity level.
+            /// </summary>
+            public int Flags;
+
+            /// <summary>
+            /// The time stamp for this message, equivalent to what GetMessageTime would return for this message.
+            /// </summary>
+            public int TimeStamp;
+
+            /// <summary>
+            /// Additional information associated with the message.
+            /// </summary>
+            public IntPtr AdditionalInformation;
+        }
+
+        public enum KeyboardState
+        {
+            KeyDown = 0x0100,
+            KeyUp = 0x0101,
+            SysKeyDown = 0x0104,
+            SysKeyUp = 0x0105
+        }
+
+        public class GlobalKeyboardHookEventArgs : HandledEventArgs
+        {
+            public KeyboardState KeyboardState { get; }
+            public LowLevelKeyboardInputEvent KeyboardData { get; }
+
+            public GlobalKeyboardHookEventArgs(
+                LowLevelKeyboardInputEvent keyboardData,
+                KeyboardState keyboardState)
+            {
+                KeyboardData = keyboardData;
+                KeyboardState = keyboardState;
+            }
+        }
+    }
+}
